Remove NSInteger cast from LTK_VALUE_IS_EVEN and LTK_VALUE_IS_ODD
- These macros no longer work agnostically with floats and integers, but they also no longer have the possibility of returning erroneous results when dealing with types incompatible with a cast to NSInteger.

+++ LTKit/Utilities/LTKMath.h
@@ -19,13 +19,13 @@
 
 #import <Foundation/Foundation.h>
 
-#define LTK_CLAMP(VALUE, MIN, MAX)	({																							\
-										__typeof__(VALUE) LTK__value = (VALUE);													\
-										__typeof__(MIN) LTK__min = (MIN);														\
-										__typeof__(MAX) LTK__max = (MAX); 														\
-										NSCAssert((LTK__min < LTK__max), @"LTK_CLAMP: MIN must be less than MAX.");				\
-										(LTK__min > LTK__value ? LTK__min : (LTK__value < LTK__max ? LTK__value : LTK__max));	\
-									})
+#define LTK_CLAMP(VALUE, MIN, MAX)			({																								\
+												__typeof__(VALUE) LTK__value = (VALUE);													\
+												__typeof__(MIN) LTK__min = (MIN);															\
+												__typeof__(MAX) LTK__max = (MAX);															\
+												NSCAssert((LTK__min < LTK__max), @"LTK_CLAMP: MIN must be less than MAX.");				\
+												(LTK__min > LTK__value ? LTK__min : (LTK__value < LTK__max ? LTK__value : LTK__max));	\
+											})
 
 #define LTK_VALUE_IS_POSITIVE(VALUE)		((VALUE) > 0)
 
@@ -35,6 +35,6 @@
 
 #define LTK_VALUE_IS_NONPOSITIVE(VALUE)	((VALUE) <= 0)
 
-#define LTK_VALUE_IS_EVEN(VALUE) 			((NSInteger)(VALUE) % 2 == 0)
+#define LTK_VALUE_IS_EVEN(VALUE) 			((VALUE) % 2 == 0)
 
-#define LTK_VALUE_IS_ODD(VALUE) 			((NSInteger)(VALUE) % 2 == 1)
+#define LTK_VALUE_IS_ODD(VALUE) 			((VALUE) % 2 == 1)
